feat(formatter): v2 of the formatter
In order to preserve comments and custom newlines, we take
a different approach. Inspired by the Dart formatter, we
will try to split the source into spans, that contain chunks,
which contain tokens. We use the AST as anchor but read the actual
tokens spans from the token stream to catch all comments.